Set dotenv plugin, moved utils/ to lib/
