+++ components/BannerWithTabs.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { breakpoints } from 'lib/theme';
-import { Text } from 'utils/ui';
+import { Text } from 'lib/ui';
 import { Box, Flex } from '@rebass/grid/emotion';
 
 interface State {
