+++ components/Header/index.tsx
@@ -7,7 +7,7 @@ import { css } from '@emotion/core';
 import { Box, Flex } from '@rebass/grid/emotion';
 
 import styled, { ThemeProps, breakpoints } from 'lib/theme';
-import { Col, Container } from 'utils/ui';
+import { Col, Container } from 'lib/ui';
 import TopNav from './top-nav';
 import Nav from './nav';
 import Cta from './cta';
