+++ components/Header/nav-item.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link';
 import { Flex, Box } from '@rebass/grid/emotion';
 import { isMobile } from 'react-device-detect';
 
-import { Col } from 'utils/ui';
+import { Col } from 'lib/ui';
 import styled, { breakpoints } from 'lib/theme';
 import SubMenu from './submenu';
