+++ lib/constants.ts
@@ -1,0 +1,1 @@
+export const graphqlEndpoint = `${process.env.MAGENTO_API}/graphql`;
