+++ package.json
@@ -28,6 +28,7 @@
 		"apollo-boost": "^0.1.23",
 		"apollo-link-context": "^1.0.12",
 		"apollo-link-http": "^1.5.9",
+		"dotenv": "^6.2.0",
 		"emotion": "^10.0.5",
 		"emotion-server": "^10.0.5",
 		"emotion-theming": "^10.0.5",
