--- utils/constants.ts
@@ -1,4 +1,0 @@
-// export const graphqlEndpoint = `${process.env.API}/graphql`;
-// TODO: Add magento api
-// export const graphqlEndpoint = 'http://localhost:4000';
-export const graphqlEndpoint = 'http://54.213.194.167/graphql';
